Update the file to its desired outcome
This is the definition of the title above which is set accidentally while I was trying to ommit the message part. Although this is a good reminder of
what to do in case you want to add a description for the commit and its subject. Now, how do I close this?